refactor: reorganize folder structure and move components
